requirements.txt
pyaudio
anyio==4.3.0
av==11.0.0
certifi==2023.7.22
charset-normalizer==3.3.2
coloredlogs==15.0.1
ctranslate2==4.1.0
docopt==0.6.2
faster-whisper==1.0.1
filelock==3.13.1
fsspec==2023.10.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
humanfriendly==10.0
idna==3.4
Jinja2==3.1.2
joblib==1.3.2
Js2Py==0.74
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.2
ollama==0.1.8
onnxruntime==1.16.2
openwakeword==0.5.1
packaging==23.2
piper-phonemize==1.1.0
piper-tts==1.2.0
protobuf==4.25.0
pyjsparser==2.7.1
PyPrind==2.11.3
pySmartDL==1.3.4
python-dotenv==1.0.0
PyYAML==6.0.1
requests==2.31.0
scipy==1.11.3
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
sympy==1.12
tflite-runtime==2.14.0
threadpoolctl==3.2.0
tokenizers==0.15.2
torch==2.1.0
torchaudio==2.1.0
tqdm==4.66.1
typing_extensions==4.8.0
tzlocal==5.2
urllib3==2.1.0
